small fixes

+++ ProjectManager.Desktop/Models/Project.cs
@@ -1,4 +1,5 @@
 using System.Collections.ObjectModel;
+using System.Windows;
 using System.Windows.Input;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
@@ -17,6 +18,11 @@ public partial class Project : ObservableObject
 
     public ICommand DeleteProjectCommand => new RelayCommand(async () =>
     {
+        var isDeleteQuestion = MessageBox.Show("Удалить проект?", "Удаление", MessageBoxButton.YesNo, MessageBoxImage.Question);
+
+        if (isDeleteQuestion != MessageBoxResult.Yes)
+            return;
+
         await ProjectService.DeleteAsync(IdProject);
     });
